requirements.txt
flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
setuptools==70.0.0
wheel==0.42.0
numpy==1.24.3
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
